[FEAT] Extract firebase
