+++ src/lib/onboardingAnalytics.ts
@@ -1,4 +1,4 @@
-import { initializeApp } from "firebase/app";
+import { app } from "./firebase";
 import {
   Analytics,
   AnalyticsCallOptions,
@@ -47,18 +47,6 @@ class OnboardingAnalytics {
   private analytics: Analytics;
 
   public constructor() {
-    const firebaseConfig = {
-      apiKey: "AIzaSyCozYr5S8ahU0WSoTS13ctjtFrleD5rZB8",
-      authDomain: "sunflower-land.firebaseapp.com",
-      projectId: "sunflower-land",
-      storageBucket: "sunflower-land.appspot.com",
-      messagingSenderId: "1061537811936",
-      appId: "1:1061537811936:web:4357cbb765c9c990f66f85",
-      measurementId: "G-EM6CNBH1F8",
-    };
-
-    // Initialize Firebase
-    const app = initializeApp(firebaseConfig);
     this.analytics = getAnalytics(app);
   }
